Conflicts:
	src/config/Makefile.config.in
	src/config/configure.in
	src/configure
	src/distributed_ls/Euclid/_hypre_Euclid.h
	src/distributed_ls/Euclid/globalObjects.c
	src/sstruct_ls/fac_amr_fcoarsen.c
	src/sstruct_ls/fac_cf_coarsen.c
	src/sstruct_ls/fac_interp2.c
	src/sstruct_ls/fac_restrict2.c
	src/sstruct_ls/maxwell_PNedelec.c
	src/sstruct_ls/maxwell_semi_interp.c
	src/sstruct_ls/node_relax.c
	src/sstruct_mv/sstruct_matrix.c
	src/struct_ls/HYPRE_struct_int.c
	src/struct_ls/cyclic_reduction.c
	src/struct_ls/pfmg.c
	src/struct_ls/pfmg2_setup_rap.c
	src/struct_ls/pfmg3_setup_rap.c
	src/struct_ls/pfmg_setup.c
	src/struct_ls/pfmg_setup_interp.c
	src/struct_ls/pfmg_setup_rap5.c
	src/struct_ls/pfmg_setup_rap7.c
	src/struct_ls/red_black_constantcoef_gs.c
	src/struct_ls/red_black_gs.c
	src/struct_ls/red_black_gs.h
	src/struct_ls/semi_setup_rap.c
	src/struct_ls/smg.c
	src/struct_ls/smg2_setup_rap.c
	src/struct_ls/smg3_setup_rap.c
	src/struct_ls/sparse_msg3_setup_rap.c
	src/struct_ls/sparse_msg_filter.c
	src/struct_mv/_hypre_struct_mv.h
	src/struct_mv/box.h
	src/struct_mv/boxloop.h
	src/struct_mv/boxloop_cuda.h
	src/struct_mv/boxloop_kokkos.h
	src/struct_mv/boxloop_raja.h
	src/struct_mv/headers
	src/struct_mv/struct_communication.c
	src/struct_mv/struct_innerprod.c
	src/struct_mv/struct_io.c
	src/struct_mv/struct_matrix.c
	src/struct_mv/struct_vector.c
	src/test/Makefile
	src/test/struct_migrate.c
	src/test/struct_newboxloop.c
	src/utilities/_hypre_utilities.h
	src/utilities/hypre_memory.h

+++ src/struct_mv/struct_innerprod.c
@@ -37,9 +37,9 @@ hypre_StructInnerProd( hypre_StructVector *x,
                    
    hypre_BoxArray  *boxes;
    hypre_Box       *box;
-    hypre_Index      loop_size;
-    hypre_IndexRef   start;
-    hypre_Index      unit_stride;
+   hypre_Index      loop_size;
+   hypre_IndexRef   start;
+   hypre_Index      unit_stride;
     
    HYPRE_Int         ndim = hypre_StructVectorNDim(x);               
    HYPRE_Int        i;
@@ -51,16 +51,16 @@ hypre_StructInnerProd( hypre_StructVector *x,
    boxes = hypre_StructGridBoxes(hypre_StructVectorGrid(y));
    hypre_ForBoxI(i, boxes)
    {
-     box   = hypre_BoxArrayBox(boxes, i);
-     start = hypre_BoxIMin(box);
+      box   = hypre_BoxArrayBox(boxes, i);
+      start = hypre_BoxIMin(box);
      
-     x_data_box = hypre_BoxArrayBox(hypre_StructVectorDataSpace(x), i);
-     y_data_box = hypre_BoxArrayBox(hypre_StructVectorDataSpace(y), i);
+      x_data_box = hypre_BoxArrayBox(hypre_StructVectorDataSpace(x), i);
+      y_data_box = hypre_BoxArrayBox(hypre_StructVectorDataSpace(y), i);
      
-     xp = hypre_StructVectorBoxData(x, i);
-     yp = hypre_StructVectorBoxData(y, i);
+      xp = hypre_StructVectorBoxData(x, i);
+      yp = hypre_StructVectorBoxData(y, i);
      
-     hypre_BoxGetSize(box, loop_size);
+      hypre_BoxGetSize(box, loop_size);
 
 #ifdef HYPRE_BOX_PRIVATE_VAR
 #undef HYPRE_BOX_PRIVATE_VAR
@@ -70,15 +70,15 @@ hypre_StructInnerProd( hypre_StructVector *x,
 #undef HYPRE_BOX_REDUCTION
 #endif
 #define HYPRE_BOX_REDUCTION reduction(+:local_result)
-     hypre_newBoxLoop2ReductionBegin(ndim, loop_size,
-				     x_data_box, start, unit_stride, xi,
-				     y_data_box, start, unit_stride, yi,local_result);
-     {
-       local_result += xp[xi] * hypre_conj(yp[yi]);		 
-     }
-     hypre_newBoxLoop2ReductionEnd(xi, yi, local_result);
+      hypre_newBoxLoop2ReductionBegin(ndim, loop_size,
+                                      x_data_box, start, unit_stride, xi,
+                                      y_data_box, start, unit_stride, yi,local_result);
+      {
+         local_result += xp[xi] * hypre_conj(yp[yi]);		 
+      }
+      hypre_newBoxLoop2ReductionEnd(xi, yi, local_result);
    }
-   process_result = (double)(local_result);
+   process_result = local_result;
    
    hypre_MPI_Allreduce(&process_result, &final_innerprod_result, 1,
                        HYPRE_MPI_REAL, hypre_MPI_SUM, hypre_StructVectorComm(x));
